few updates

+++ NightToday/losses.py
@@ -690,7 +690,7 @@ class SharpFusionLoss(torch.nn.Module):
             if I_vi.shape[1] == 3 else I_vi * 0.5 + 0.5
         I_ir = (0.299 * I_ir[:, 0:1] + 0.587 * I_ir[:, 1:2] + 0.114 * I_ir[:, 2:3]) * 0.5 + 0.5 \
             if I_ir.shape[1] == 3 else I_ir * 0.5 + 0.5
-        mask_hl = erosion(I_vi < 0.9, kernel=torch.ones(5, 5, device=I_vi.device))
+        mask_hl = erosion((I_vi < 0.9).float(), kernel=torch.ones(5, 5, device=I_vi.device))
         # -------- Gradient Loss --------
         G_f = sobel(I_f).abs()
         G_vi = sobel(I_vi).abs() * mask_hl  # focus on dark areas where thermal has more details
